Added Vehicle Lock function for the AI
Locks empty spots in AI vehicles to prevent AI from gettin in when not
supposed to, only allowing ones that are determined by the mission maker
to get in (currently only the squad leader but at a later point will be
an integral part of the transport script). This prevents AI squads from
running many killometers back to base to embark a vehicle.